+++ TP_LABS/App.config
@@ -1,0 +1,23 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+    <configSections>
+        <section name="nlog" type="NLog.Config.ConfigSectionHandler, NLog" />
+    </configSections>
+  
+    <startup> 
+        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6.1" />
+    </startup>
+
+    <nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+      <targets>
+         <target name="logfile" xsi:type="File" fileName="C:\\Users\\user\\Desktop\\logfile.txt" />
+         <target name="logfile1" xsi:type="File" fileName="C:\\Users\\user\\Desktop\\ex.txt" 
+               />
+      </targets>
+      <rules>
+         <logger name="*" minlevel="Info" writeTo="logfile" />
+         <logger name="*" minlevel="Error" writeTo="logfile1" />
+      </rules>
+    </nlog>
+
+</configuration>
