+++ TP_LABS/ParkingNotFoundException.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TP_LABS
+{
+    public class ParkingNotFoundException : Exception
+    {
+        public ParkingNotFoundException(int i) : base("Не найден автомобиль по месту " + i) { }
+    }
+}
