+++ TP_LABS/ParkingOccupiedPlaceException.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TP_LABS
+{
+    public class ParkingOccupiedPlaceException : Exception
+    {
+        public ParkingOccupiedPlaceException(int i) : base("На месте " + i + " уже стоит автомобиль") { }
+    }
+}
