+++ TP_LABS/ParkingOverflowException.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TP_LABS
+{
+    public class ParkingOverflowException : Exception
+    {
+        public ParkingOverflowException() : base("На парковке нет свободных мест") { }
+    }
+}
