--- TP_LABS/TP_LABS.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.779
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TP_LABS", "TP_LABS\TP_LABS.csproj", "{BB88961F-861A-447C-8C1E-AD6189ADE278}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{BB88961F-861A-447C-8C1E-AD6189ADE278}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BB88961F-861A-447C-8C1E-AD6189ADE278}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BB88961F-861A-447C-8C1E-AD6189ADE278}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BB88961F-861A-447C-8C1E-AD6189ADE278}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D1C9BB55-4BD4-4EAC-BF7E-4FB0F3D34A2F}
-	EndGlobalSection
-EndGlobal
