+++ TP_LABS/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="NLog" version="4.6.8" targetFramework="net461" />
+</packages>
